Version 1.9: Alpha From both notification panel and notification list redirection has been implemented.

+++ app/src/main/java/in/gov/mahapocra/mahavistaarai/ui/screens/notification/NotificationActivity.kt
@@ -65,7 +65,7 @@ class NotificationActivity : AppCompatActivity() {
                     adapter = NotificationAdapter(notificationJsonArray) { jsonObject ->
                         startActivity(Intent(this@NotificationActivity,
                             DetailedNotificationActivity::class.java).apply {
-                                putExtra("id", jsonObject.optInt("id"))
+                                putExtra("id", jsonObject.optLong("id"))
                         })
                     }
                 }
